fix: update shortcut key bindings and improve sorting utility
- Changed the shortcut for toggling Zen mode to use "Shift+$mod+Z".
- Removed redundant sorting functions from SettingShortcuts and moved sorting logic to utils.
- Enhanced the sorting utility for shortcut keys to ensure consistent ordering of modifier keys.
- Updated GlobalHotkeysProvider to utilize the new sorting logic and added a check for normal layer focus.
